simplify implementation

+++ LeetCodeSolutions/SingleNumber/SingleNumber.cs
@@ -3,11 +3,11 @@ public class SingleNumberSolution
 {
     public int SingleNumber(int[] nums)
     {
-        uint outlier = 0b0;
-        foreach (uint num in nums)
+        int outlier = 0b0;
+        foreach (int num in nums)
         {
             outlier = num ^ outlier;
         }
-        return (int)outlier;
+        return outlier;
     }
 }
